let users examine info for any game they have had
i will use this to check if i am actually updating the back end, since it is
using a get, i will likely also leave it in a bonus feature.

it required investigating what kind of objects the AJAX calls were returning.
the key work was done in the ui.js file.